Move DuplicateEntryCountAttribute to the correct namespace
