+++ SAGESharp/SLB/Emitter.cs
@@ -4,6 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 using SAGESharp.IO;
+using SAGESharp.IO.Binary;
 using System.Collections.Generic;
 
 namespace SAGESharp.SLB
